add support u2.connect_usb(serial, init=False)
